docs: update

+++ examples/0_embedding/train_llm.py
@@ -99,7 +99,7 @@ class TrainingArguments(transformers.TrainingArguments):
     normalized: bool = field(default=True)
     use_inbatch_neg: bool = field(default=True, metadata={"help": "Freeze the parameters of position embeddings"})
     gradient_accumulation_steps: int = field(default=1024)
-    fp16: bool = field(default=True)
+    bf16: bool = field(default=True)
     logging_steps: int = field(default=100)
